remove catch TODO. This method is for preparation only.
The prepareException(e) method just repackages a given Exception and
might extract additional information. This is useful if e.g. a JPA
provider wraps a Database Exception in his own or for adding info
about failed tests in a ConstraintViolationException
which by default has a meaningless text.

We don't do catch integration in this spot because we don't throw
anything.