removed obsolete TODO for Weld
This scenario must not happen. The ClassLoader used to load
the Extension is also the one used at shutdown. Even in EAR
scenarios with a hierarchic setup we will hopefully get the
correct TCCL. The mem will get freed up lastest after the EAR
gets undeployed anyway. Sparse re-deployments of portions of
an EAR are not guaranteed to work by the EE spec neither.