Fixes DELTASPIKE-400 (null beans in InjectionPoints)
I've added a couple of tests now, we can start building on them for BeanBuilder.

The BaseImmutableBean now checks for the Bean in InjectionPoints and uses the new InjectionPointWrapper with self as the bean if there is no bean set on the InjectionPoint.

I've not tested to see if everything is wired up correctly, but all the tests are passing.