DELTASPIKE-1402 dynamic reload of PropertyFileConfigSource
DeltaSpike can now dynamically reload configuration of property files in a file:// location.
The time after which we look for file modification (via lastModified time stamp of the file)
can be configured with a 'deltaspike_reload=60' property in seconds.
In this case 60 seconds.
If this property is not set we check for changes every 300 seconds or 5 Minutes.