DELTASPIKE-1426: Improve ProxyFactory performance on start
Class.getMethond(name, args) is expensive operation, that is used
intensively on start, when DeltaspikeProxyFactory generates new proxies.
It was used even for checks of public abstract
methods of proxy's base class, but this methods cannot be overridden.
This patch collects all leave methods and skips them in the check.